twister: split hardwaremap class out
Split hardware map class into own file.
